week_3/azure_bot/requirements.txt
botbuilder-integration-aiohttp>=4.15.0
azure-ai-textanalytics
